+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Delta tables metadata compatibility when name and/or description is missing
+- Delta tables history compatibility when created with Spark
 
 ## [0.1.0] - 2025-02-15
 Initial release of `laketower`.