+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.0] - 2025-02-25
+Introducing the Laketower web application!
+
 ### Added
 - `web` module
     - List all registered tables
@@ -33,5 +37,6 @@ Initial release of `laketower`.
     - View a given table with simple query builder
     - Query all registered tables with DuckDB SQL dialect
 
-[Unreleased]: https://github.com/datalpia/laketower/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/datalpia/laketower/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/datalpia/laketower/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/datalpia/laketower/releases/tag/0.1.0